Ensure that containers are restart after the docker restarts
If the module decides to reload the docker daemon then all containers are stopped,
this change ensures that all containers are started after the refresh of the service.

The functionality can be disabled with the restart_service_on_docker_refresh parameter in docker::run.
For it to work the docker_service parameter needs to be set, indicating that the docker daemon is managed by this module.

Change-Id: Icf2213942ddf4d7ef871ef8c308f1e3c4d7c1f00